Refactor useMeshBounds to regular meshBounds function, show usage in a plugin
