--- apps/docs/src/content/reference/extras/useMeshBounds.mdx
@@ -1,32 +1,0 @@
----
-order: 6.8
-category: '@threlte/extras'
-name: 'useMeshBounds'
-type: 'hook'
-sourcePath: 'packages/extras/src/lib/hooks/useMeshBounds.ts'
----
-
-A raycast function that will first check for collision with the object's bounding sphere, then it will check against the object's bounding box but **only** if the bounding box has been previously computed. You can use this function if you want to trade pointer precision for an increase in performance.
-
-## Usage
-
-`useMeshBounds` returns a function that you can pass to an object's [`raycast`](https://threejs.org/docs/index.html?q=mesh#api/en/objects/Mesh.raycast) prop.
-
-```svelte
-<script>
-  import { useMeshBounds } from '@threlte/extras'
-
-  const meshBounds = useMeshBounds()
-</script>
-
-<T.Mesh raycast={meshBounds}>
-  <T.BoxGeometry />
-</T.Mesh>
-```
-
-<Tip type="tip">
-  If your meshes or models aren't very complex in terms of geometry, `useMeshBounds` won't provide
-  much of a performance boost.
-</Tip>
-
-<Example path="extras/use-mesh-bounds" />
